Split out PlatformWebView on WinForms.
